aggregator/requirements.txt
Flask==2.3.3
flask_cors==4.0.0
sqlalchemy==2.0.20
python-dotenv==1.0.0
